pyproject.toml
[project]
name = "api-getaway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "google-genai>=1.31.0",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "sqlalchemy-utils>=0.41.2",
    "tiktoken>=0.11.0",
    "uvicorn>=0.35.0",
]
